pyproject.toml
[build-system]
requires = ["setuptools>=68.0,<78.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policydhara"
version = "0.1.0"
description = "Python library and CLI for tracking Indian development policies across 21 sectors and 300+ sources"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ImpactMojo", email = "hello@impactmojo.in"},
]
keywords = [
    "india", "policy", "government", "tracker", "development",
    "legislation", "parliament", "sectors", "governance",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "beautifulsoup4>=4.11",
    "lxml>=4.9",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://varnasr.github.io/PolicyDhara"
Repository = "https://github.com/Varnasr/PolicyDhara"
Issues = "https://github.com/Varnasr/PolicyDhara/issues"

[project.scripts]
policydhara = "policydhara.cli:main"

[tool.setuptools.packages.find]
include = ["policydhara*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
